requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cupy-cuda12x==12.3.0
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastrlock==0.8.2
filelock==3.13.3
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
langcodes==3.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
preshed==3.0.9
pydantic==2.6.4
pydantic_core==2.16.3
pytesseract==0.3.10
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sacremoses==0.1.1
safetensors==0.4.2
sentencepiece==0.2.0
setuptools==69.2.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
tokenizers==0.15.2
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
wheel==0.43.0
